Classify landcover for SpatRaster object

Prepare for fitting SDMs by reclassifying an existing landscape raster according to the classifications used by a specific set of species distribution models (SDM).

Usage

# S3 method for class 'SpatRaster'
classify_landcover(x, SDM, coltab = TRUE, verbose = TRUE, ...)

Arguments

x

SpatRaster

SDM

The name of intended species distribution model: "riparian", "waterbird_fall", "waterbird_win", or "tima"

coltab

logical; if TRUE add default color palette

verbose

logical; if TRUE then print details associated with warning messages

...

Unused

Value

SpatRaster with the same number of layers as the input x

Details

This function is called by create_predictor_stack() and is not generally intended to be called directly. It calls on internal datasets to crosswalk from land cover classes listed in the key to the predictors expected by the selected SDM group. The input raster should already be encoded with the land cover classes listed in the key. To help with creating such a raster, see classify_landcover.sf() to map land cover polygons to the land cover classes in the key.

A warning is given if there are land cover classes present in the landscape that do not map to any of the predictors for the selected SDM group. This warnings may represent significant problems for fitting SDMs and should be carefully reviewed. It is recommended to review the corresponding internal datasets (predictors_riparian, predictors_waterbird_fall, predictors_waterbird_win, or predictors_tima) for the list of expected predictors and how they map to land cover classes in the key. Check whether the selected SDM group expects more specific land cover classes or subclasses; the input raster may need to be reclassified before proceeding.

Examples

r <- terra::rast(matrix(sample(c(11,19,71,72,90), size = 100, replace = TRUE),
         ncol = 10, nrow = 10))
r <- suppressWarnings(classify_landcover(r, SDM = 'riparian'))