Land cover classification scheme for the DeltaMultipleBenefits framework

Major land cover classes and subclasses designed to work with the existing metrics and species distribution models within the DeltaMultipleBenefits framework. It includes both natural and agricultural land cover classes, and is organized hierarchically into major land cover classes and subclasses. Also included are default values for labels and color coding used in maps.

Usage

key

Format

key A data frame with 41 rows and 7 columns:

CODE_BASELINE

Numeric value used to encode rasters

CODE_NAME

Text string joining major land cover classes to subclasses with a '_'

CLASS

Major land cover class grouping

SUBCLASS

Land cover subclass

DETAIL

Further land cover detail

LABEL

Default value for labels used in maps and plots

COLOR

Default hex color code used in maps

Source

Dybala et al. In prep (URL to be determined)