Update waterbird predictors: covertype

Helper function for updating covertype predictors for the waterbird distribution models.

Usage

update_covertype(
  landscape,
  key,
  SDM,
  mask = NULL,
  pathout,
  landscape_name,
  overwrite = FALSE
)

Arguments

landscape

SpatRaster created by terra::rast()

key

tibble, dataframe, or character string defining a filepath passed to readr::read_csv(), used to interpret the raster values in landscape to land cover class names; see Details

SDM

The name of intended species distribution model: "waterbird_fall", or "waterbird_win"

mask

Optional filepath to a raster that should be used to mask the output, e.g. a study area boundary

pathout, landscape_name

Character strings defining the filepath (pathout/SDM/landscape_name) where output rasters should be written

overwrite

Logical; passed to terra::writeRaster(); default FALSE

Value

Nothing; all files written to pathout

Details

Classifies the landscape rasters according to the land cover classes that were originally surveyed, which are the only classes for which predictions should be generated from the waterbird distribution models. Generates file covertype.tif at location pathout/SDM/scenario_name/.

See also

Examples

# See vignette